同步activiti修改

+++ modules/activiti-engine/src/test/java/org/activiti/engine/test/api/task/TaskServiceTest.java
@@ -32,6 +32,7 @@ import org.activiti.engine.identity.Group;
 import org.activiti.engine.identity.User;
 import org.activiti.engine.impl.TaskServiceImpl;
 import org.activiti.engine.impl.history.HistoryLevel;
+import org.activiti.engine.impl.persistence.entity.CommentEntity;
 import org.activiti.engine.impl.persistence.entity.HistoricDetailVariableInstanceUpdateEntity;
 import org.activiti.engine.impl.test.PluggableActivitiTestCase;
 import org.activiti.engine.runtime.ProcessInstance;
@@ -148,6 +149,48 @@ public class TaskServiceTest extends PluggableActivitiTestCase {
       taskService.deleteTask(taskId, true);
     }
   }
+  
+  public void testCustomTaskComments() {
+    if (processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.ACTIVITY)) {
+      Task task = taskService.newTask();
+      task.setOwner("johndoe");
+      taskService.saveTask(task);
+      String taskId = task.getId();
+      
+      identityService.setAuthenticatedUserId("johndoe");
+      String customType1 = "Type1";
+      String customType2 = "Type2";
+      
+      Comment comment = taskService.addComment(taskId, null, "This is a regular comment");
+      Comment customComment1 = taskService.addComment(taskId, null, customType1, "This is a custom comment of type Type1");
+      Comment customComment2 = taskService.addComment(taskId, null, customType1, "This is another Type1 comment");
+      Comment customComment3 = taskService.addComment(taskId, null, customType2, "This is another custom comment. Type2 this time!");
+      
+      assertEquals(CommentEntity.TYPE_COMMENT, comment.getType());
+      assertEquals(customType1, customComment1.getType());
+      assertEquals(customType2, customComment3.getType());
+      
+      assertNotNull(taskService.getComment(comment.getId()));
+      assertNotNull(taskService.getComment(customComment1.getId()));
+      
+      List<Comment> regularComments = taskService.getTaskComments(taskId);
+      assertEquals(1, regularComments.size());
+      assertEquals("This is a regular comment", regularComments.get(0).getFullMessage());
+      
+      List<Event> allComments = taskService.getTaskEvents(taskId);
+      assertEquals(4, allComments.size());
+      
+      List<Comment> type2Comments = taskService.getCommentsByType(customType2);
+      assertEquals(1, type2Comments.size());
+      assertEquals("This is another custom comment. Type2 this time!", type2Comments.get(0).getFullMessage());
+      
+      List<Comment> taskTypeComments = taskService.getTaskComments(taskId, customType1);
+      assertEquals(2, taskTypeComments.size());
+      
+      // Clean up
+      taskService.deleteTask(taskId, true);
+    }
+  }
 
   public void testTaskAttachments() {
     if (processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.ACTIVITY)) {
@@ -166,6 +209,34 @@ public class TaskServiceTest extends PluggableActivitiTestCase {
       assertNull(attachment.getProcessInstanceId());
       assertEquals("http://weather.com", attachment.getUrl());
       assertNull(taskService.getAttachmentContent(attachment.getId()));
+      
+      // Finally, clean up
+      taskService.deleteTask(taskId);
+      
+      assertEquals(0, taskService.getTaskComments(taskId).size());
+      assertEquals(1, historyService.createHistoricTaskInstanceQuery().taskId(taskId).list().size());
+
+      taskService.deleteTask(taskId, true);
+    }
+  }
+  
+  public void testSaveTaskAttachment() {
+    if (processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.ACTIVITY)) {
+      Task task = taskService.newTask();
+      task.setOwner("johndoe");
+      taskService.saveTask(task);
+      String taskId = task.getId();
+      identityService.setAuthenticatedUserId("johndoe");
+      
+      // Fetch attachment and update its name
+      taskService.createAttachment("web page", taskId, null, "weatherforcast", "temperatures and more", "http://weather.com");
+      Attachment attachment = taskService.getTaskAttachments(taskId).get(0);
+      attachment.setName("UpdatedName");
+      taskService.saveAttachment(attachment);
+      
+      // Refetch and verify
+      attachment = taskService.getTaskAttachments(taskId).get(0);
+      assertEquals("UpdatedName", attachment.getName());
       
       // Finally, clean up
       taskService.deleteTask(taskId);
@@ -420,7 +491,7 @@ public class TaskServiceTest extends PluggableActivitiTestCase {
     assertEquals(user.getId(), task.getAssignee());
     
     // Unclaim the task
-    taskService.claim(task.getId(), null);
+    taskService.unclaim(task.getId());
     
     task = taskService.createTaskQuery().taskId(task.getId()).singleResult();
     assertNull(task.getAssignee());
@@ -523,6 +594,55 @@ public class TaskServiceTest extends PluggableActivitiTestCase {
     Map<String, Object> variables = runtimeService.getVariables(processInstance.getId());
     assertEquals(1, variables.size());
     assertEquals("myValue", variables.get("myParam"));
+  }
+  
+  @Deployment(resources = { 
+  "org/activiti/engine/test/api/twoTasksProcess.bpmn20.xml" })
+public void testCompleteWithParametersTask2() {
+  ProcessInstance processInstance = runtimeService.startProcessInstanceByKey("twoTasksProcess");
+
+  // Fetch first task
+  Task task = taskService.createTaskQuery().singleResult();
+  assertEquals("First task", task.getName());
+
+  // Complete first task
+  Map<String, Object> taskParams = new HashMap<String, Object>();
+  taskParams.put("myParam", "myValue");
+  taskService.complete(task.getId(), taskParams, false); // Only difference with previous test
+
+  // Fetch second task
+  task = taskService.createTaskQuery().singleResult();
+  assertEquals("Second task", task.getName());
+
+  // Verify task parameters set on execution
+  Map<String, Object> variables = runtimeService.getVariables(processInstance.getId());
+  assertEquals(1, variables.size());
+  assertEquals("myValue", variables.get("myParam"));
+}
+  
+  @Deployment
+  public void testCompleteWithTaskLocalParameters() {
+  	ProcessInstance processInstance = runtimeService.startProcessInstanceByKey("testTaskLocalVars");
+
+    // Fetch first task
+    Task task = taskService.createTaskQuery().singleResult();
+
+    // Complete first task
+    Map<String, Object> taskParams = new HashMap<String, Object>();
+    taskParams.put("a", 1);
+    taskParams.put("b", 1);
+    taskService.complete(task.getId(), taskParams, true);
+    
+    // Verify vars are not stored process instance wide
+    assertNull(runtimeService.getVariable(processInstance.getId(), "a"));
+    assertNull(runtimeService.getVariable(processInstance.getId(), "b"));
+    
+    // verify script listener has done its job
+    assertEquals(new Integer(2), runtimeService.getVariable(processInstance.getId(), "sum"));
+
+    // Fetch second task
+    task = taskService.createTaskQuery().singleResult();
+
   }
   
   public void testSetAssignee() {
@@ -1152,7 +1272,89 @@ public class TaskServiceTest extends PluggableActivitiTestCase {
       
     }
   }
-  
+
+  public void testResolveTaskNullTaskId() {
+    try {
+      taskService.resolveTask(null);
+      fail();
+    } catch (ActivitiException ae) {
+      assertTextPresent("taskId is null", ae.getMessage());
+    }
+  }
+
+  public void testResolveTaskUnexistingTaskId() {
+    try {
+      taskService.resolveTask("blergh");
+      fail();
+    } catch (ActivitiException ae) {
+      assertTextPresent("Cannot find task with id", ae.getMessage());
+    }
+  }
+
+  public void testResolveTaskWithParametersNullParameters() {
+    Task task = taskService.newTask();
+    task.setDelegationState(DelegationState.PENDING);
+    taskService.saveTask(task);
+
+    String taskId = task.getId();
+    taskService.resolveTask(taskId, null);
+
+    if (processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.AUDIT)) {
+      historyService.deleteHistoricTaskInstance(taskId);
+    }
+
+    // Fetch the task again
+    task = taskService.createTaskQuery().taskId(taskId).singleResult();
+    assertEquals(DelegationState.RESOLVED, task.getDelegationState());
+
+    taskService.deleteTask(taskId, true);
+  }
+
+  @SuppressWarnings("unchecked")
+  public void testResolveTaskWithParametersEmptyParameters() {
+    Task task = taskService.newTask();
+    task.setDelegationState(DelegationState.PENDING);
+    taskService.saveTask(task);
+
+    String taskId = task.getId();
+    taskService.resolveTask(taskId, Collections.EMPTY_MAP);
+
+    if (processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.AUDIT)) {
+      historyService.deleteHistoricTaskInstance(taskId);
+    }
+
+    // Fetch the task again
+    task = taskService.createTaskQuery().taskId(taskId).singleResult();
+    assertEquals(DelegationState.RESOLVED, task.getDelegationState());
+
+    taskService.deleteTask(taskId, true);
+  }
+
+  @Deployment(resources = { "org/activiti/engine/test/api/twoTasksProcess.bpmn20.xml" })
+  public void testResolveWithParametersTask() {
+    ProcessInstance processInstance = runtimeService.startProcessInstanceByKey("twoTasksProcess");
+
+    // Fetch first task
+    Task task = taskService.createTaskQuery().singleResult();
+    assertEquals("First task", task.getName());
+
+    task.delegate("johndoe");
+
+    // Resolve first task
+    Map<String, Object> taskParams = new HashMap<String, Object>();
+    taskParams.put("myParam", "myValue");
+    taskService.resolveTask(task.getId(), taskParams);
+
+    // Verify that task is resolved
+    task = taskService.createTaskQuery().taskDelegationState(DelegationState.RESOLVED).singleResult();
+    assertEquals("First task", task.getName());
+
+    // Verify task parameters set on execution
+    Map<String, Object> variables = runtimeService.getVariables(processInstance.getId());
+    assertEquals(1, variables.size());
+    assertEquals("myValue", variables.get("myParam"));
+  }
+
   @Deployment(resources = { "org/activiti/engine/test/api/oneTaskProcess.bpmn20.xml" })
   public void testDeleteTaskPartOfProcess() {
     runtimeService.startProcessInstanceByKey("oneTaskProcess");
